refactor: use UnitNotFound error from application domain
There is no need for each domain to have a version of the same error.
Use UnitNotFound from the application errors package rather than a local
copy. Alls for the unitstate errors package to be removed.
